Moved game engine setup to useGameEngine hook
